Huge overhaul on Booking Service and Controller

+++ backend/src/main/java/ca/mcgill/ecse321/Mar1HotelSystem/service/BookingService.java
@@ -1,19 +1,21 @@
 package ca.mcgill.ecse321.Mar1HotelSystem.service;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import ca.mcgill.ecse321.Mar1HotelSystem.dao.BookingRepository;
+import ca.mcgill.ecse321.Mar1HotelSystem.dao.GeneralUserRepository;
+import ca.mcgill.ecse321.Mar1HotelSystem.dao.PaymentRepository;
+import ca.mcgill.ecse321.Mar1HotelSystem.dao.RoomRepository;
 import ca.mcgill.ecse321.Mar1HotelSystem.exception.Mar1HotelSystemException;
 import ca.mcgill.ecse321.Mar1HotelSystem.model.Booking;
 import ca.mcgill.ecse321.Mar1HotelSystem.model.GeneralUser;
 import ca.mcgill.ecse321.Mar1HotelSystem.model.Payment;
 import ca.mcgill.ecse321.Mar1HotelSystem.model.Room;
 
+import java.util.List;
 /**
  * Service class to manage Booking Entities
  * 
@@ -25,17 +27,40 @@ public class BookingService {
     @Autowired
     BookingRepository bookingRepository;
 
+    @Autowired
+    GeneralUserRepository generalUserRepository;
+
+    @Autowired
+    RoomRepository roomRepository;
+
+    @Autowired
+    PaymentRepository paymentRepository;
+
     /**
      * Service method to create and save a new booking.
      */
     @Transactional
-    public Booking createBooking(Payment payment, GeneralUser generalUser, Room room) {
+    public Booking createBooking(String generalUserEmail, int roomId, int paymentId) {
         Booking booking = new Booking();
-        booking.setPayment(payment);
-        booking.setGeneralUser(generalUser);
-        booking.setRoom(room);
-        if(payment == null || generalUser.getFirstName() == ""  || generalUser.getLastName() == "" || generalUser.getEmail() == "" || generalUser == null) {
-            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, "Booking must have a payment, a user and a room that all follow convention.");
+        String errorCreate = "On createBooking: ";
+
+        try {
+            Payment payment = paymentRepository.findPaymentByPaymentId(paymentId);
+            booking.setPayment(payment);
+        } catch (Exception e) {
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, errorCreate + "Payment with id " + paymentId + " does not exist.");
+        }
+        try {
+            GeneralUser generalUser = generalUserRepository.findGeneralUserByEmail(generalUserEmail);
+            booking.setGeneralUser(generalUser);
+        } catch (Exception e) {
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, errorCreate + "General user with email " + generalUserEmail + " does not exist.");
+        }
+        try {
+            Room room = roomRepository.findRoomByRoomId(roomId);
+            booking.setRoom(room);
+        } catch (Exception e) {
+            throw new Mar1HotelSystemException(HttpStatus.BAD_REQUEST, errorCreate + "Room with id " + roomId + " does not exist.");
         }
         bookingRepository.save(booking);
         return booking;
@@ -45,20 +70,48 @@ public class BookingService {
      * Service method to update an existing booking.
      */
     @Transactional
-    public Booking updateBooking(Booking booking) {
-        return bookingRepository.save(booking);
+    public Booking updateBooking(int bookingId, String generalUserEmail, int roomId, int paymentId) {
+        Booking bookingToUpdate = bookingRepository.findBookingByBookingId(bookingId);
+        String errorUpdate = "On updateBooking: ";
+
+        if(bookingToUpdate == null) {
+            throw new Mar1HotelSystemException(HttpStatus.NOT_FOUND, errorUpdate + "Booking with id " + bookingId + " does not exist.");
+        }
+
+        Payment payment = paymentRepository.findPaymentByPaymentId(paymentId);
+        if (payment == null) {
+            throw new Mar1HotelSystemException(HttpStatus.NOT_FOUND, errorUpdate + "Payment with id " + paymentId + " does not exist.");
+        }
+        GeneralUser generalUser = generalUserRepository.findGeneralUserByEmail(generalUserEmail);
+        if (generalUser == null) {
+            throw new Mar1HotelSystemException(HttpStatus.NOT_FOUND, errorUpdate + "General user with email " + generalUserEmail + " does not exist.");
+        }
+        
+        Room room = roomRepository.findRoomByRoomId(roomId);
+        if (room == null) {
+            throw new Mar1HotelSystemException(HttpStatus.NOT_FOUND, errorUpdate + "Room with id " + roomId + " does not exist.");
+        }
+        
+        bookingToUpdate.setPayment(payment);
+        bookingToUpdate.setGeneralUser(generalUser);
+        bookingToUpdate.setRoom(room);
+
+        return bookingRepository.save(bookingToUpdate);
     }
 
     /**
-     * Service method to delete a booking. 
+     * Service method to delete a booking.
      */
     @Transactional
-    public void deleteBooking(int bookingId) {
-        Booking booking = this.getBookingById(bookingId);
+    public Booking deleteBooking(int bookingId) {
+        Booking booking = bookingRepository.findBookingByBookingId(bookingId);
+        if (booking == null) {
+            throw new Mar1HotelSystemException(HttpStatus.NOT_FOUND,
+                    "Can't delete: Booking with id " + bookingId + " does not exist.");
+        }
         bookingRepository.delete(booking);
-
+        return booking;
     };
-    
 
     /**
      * Service method to retrieve a booking by ID.
@@ -66,8 +119,9 @@ public class BookingService {
     @Transactional
     public Booking getBookingById(int bookingId) {
         Booking booking = bookingRepository.findBookingByBookingId(bookingId);
-        if(booking == null) {
-            throw new Mar1HotelSystemException(HttpStatus.NOT_FOUND, "Booking with id " + bookingId + " does not exist.");
+        if (booking == null) {
+            throw new Mar1HotelSystemException(HttpStatus.NOT_FOUND,
+                    "Can't get: Booking with id " + bookingId + " does not exist.");
         }
         return booking;
     }
@@ -76,7 +130,7 @@ public class BookingService {
      * Service method to retrieve all bookings.
      */
     @Transactional
-    public Iterable<Booking> getAllBookings() {
-        return bookingRepository.findAll();
+    public List<Booking> getAllBookings() {
+        return ServiceUtils.toList(bookingRepository.findAll());
     }
-  }
+}
